backend/requirements.txt
Django==4.2.*
djangorestframework==3.14.*
celery==5.3.*
redis==5.0.*
mysqlclient==2.2.*
psycopg[binary]==3.1.*
django-environ==0.11.*
django-cors-headers==3.14.*
drf-spectacular==0.27.*
gunicorn==21.2.*
PyYAML==6.0.*
requests==2.31.*
PyJWT==2.8.*
python-json-logger==2.0.7
grpcio==1.62.*
grpcio-tools==1.62.*
ldap3==2.9.*
pytest==7.4.*
pytest-django==4.7.*
pytest-cov==4.1.*
pytest-mock==3.15.*
markdown==3.5.*
python-docx==1.0.*
